Fixed Refactor Classes
