+++ app/src/main/java/com/mobiles/senecard/activitiesBusinessOwner/activityBusinessOwnerEditLoyalty/ActivityBusinessOwnerEditLoyalty.kt
@@ -1,70 +1,29 @@
 package com.mobiles.senecard.activitiesBusinessOwner.activityBusinessOwnerEditLoyalty
 
-import android.content.Intent
 import android.os.Bundle
-import android.widget.Toast
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.mobiles.senecard.activitiesBusinessOwner.activityBusinessOwnerAdvertisementsCreate.ActivityBusinessOwnerAdvertisementsCreate
-import com.mobiles.senecard.adapters.AdvertisementAdapter
-import com.mobiles.senecard.databinding.ActivityBusinessOwnerAdvertisementsBinding
-import com.mobiles.senecard.model.entities.Advertisement
+import com.mobiles.senecard.activitiesBusinessOwner.activityBusinessOwnerEditProfile.ViewModelBusinessOwnerEditProfile
 
 class ActivityBusinessOwnerEditLoyalty : AppCompatActivity() {
 
-    private lateinit var binding: ActivityBusinessOwnerAdvertisementsBinding
-    private lateinit var adapter: AdvertisementAdapter
-    private var advertisements: MutableList<Advertisement> = mutableListOf()
-    private var businessOwnerId: String? = null
-    private var storeId: String? = null
-    private var storeName: String? = null
+    private lateinit var binding: ActivityBusinessOwnerEditLoyalty
+    private val viewModel: ViewModelBusinessOwnerEditLoyalty by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        // Initialize the binding
-        binding = ActivityBusinessOwnerAdvertisementsBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-
-        // Sample advertisement data (for testing purposes)
-        advertisements.addAll(
-            listOf(
-                Advertisement("1", "Store1", "Burrito Special", "20% OFF ALL BURRITOS", null, "01/01/2024", "01/31/2024", true),
-                Advertisement("2", "Store2", "Drink Deal", "2x1 DRINKS", null, "01/01/2024", "01/31/2024", true)
-            )
-        )
-
-        // Initialize the adapter with delete callback
-        adapter = AdvertisementAdapter(advertisements) { advertisementToDelete ->
-            deleteAdvertisement(advertisementToDelete)
-        }
-
-        // Set up RecyclerView
-        binding.rvAdvertisements.layoutManager = LinearLayoutManager(this)
-        binding.rvAdvertisements.adapter = adapter
-
-        // Handle back button click
-        binding.btnBack.setOnClickListener {
-            finish()
-        }
-
-        // Handle navigation to the create advertisement view
-        binding.btnAdd.setOnClickListener {
-            val intent = Intent(this, ActivityBusinessOwnerAdvertisementsCreate::class.java)
-            intent.putExtra("businessOwnerId", businessOwnerId)
-            intent.putExtra("storeId", storeId)
-            intent.putExtra("storeName", storeName)
-            startActivity(intent)
-        }
+        setupBinding()
+        setupObservers()
     }
 
-    private fun deleteAdvertisement(advertisement: Advertisement) {
-        // Remove the advertisement from the list
-        advertisements.remove(advertisement)
-        adapter.updateAdvertisements(advertisements)
-
-        // Optional: Perform additional actions like updating the database
-        Toast.makeText(this, "${advertisement.title} deleted", Toast.LENGTH_SHORT).show()
+    private fun setupBinding() {
     }
+
+    private fun setupObservers() {
+        // Observe user data or other relevant data here
+
+        // Navigation function
+    }
+
 }
