--- app/src/main/java/com/mobiles/senecard/activitiesBusinessOwner/activityBusinessOwnerEditProfile/NavigationDestination.kt
@@ -1,7 +1,0 @@
-package com.mobiles.senecard.activitiesBusinessOwner.activityBusinessOwnerEditProfile
-
-enum class NavigationDestination {
-    ADVERTISEMENTS,
-    QR_SCANNER,
-    INITIAL
-}
